Exceptions done

+++ src/main/java/com/willrsousa/crudchallenge/services/ClientService.java
@@ -3,6 +3,8 @@ package com.willrsousa.crudchallenge.services;
 import com.willrsousa.crudchallenge.dto.ClientDTO;
 import com.willrsousa.crudchallenge.entities.Client;
 import com.willrsousa.crudchallenge.repositories.ClientRepository;
+import com.willrsousa.crudchallenge.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -17,7 +19,7 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public ClientDTO findById(Long id){
-        Client client = clientRepository.findById(id).get();
+        Client client = clientRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Cliente não encontrado"));
         return new ClientDTO(client);
     }
 
@@ -29,22 +31,31 @@ public class ClientService {
 
     @Transactional
     public ClientDTO save(ClientDTO dto){
-        Client client = new Client();
-        dtoToEntity(dto,client);
-        client = clientRepository.save(client);
-        return new ClientDTO(client);
+            Client client = new Client();
+            dtoToEntity(dto,client);
+            client = clientRepository.save(client);
+            return new ClientDTO(client);
+
     }
 
     @Transactional
     public ClientDTO update(Long id, ClientDTO dto){
+        try{
         Client client = clientRepository.getReferenceById(id);
         dtoToEntity(dto, client);
         client = clientRepository.save(client);
         return new ClientDTO(client);
+        }
+        catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Cliente não encontrado!");
+        }
     }
 
     @Transactional
     public void delete(Long id){
+        if(!clientRepository.existsById(id)){
+            throw new ResourceNotFoundException("Cliente não encontrado!");
+        }
         clientRepository.deleteById(id);
     }
